refactor: rename http.client module to httpRequest
